+++ src/index.js
@@ -94,6 +94,7 @@ import TokensController from './controllers/tokens.js';
 import ImsOrgAccessController from './controllers/ims-org-access.js';
 import FeatureFlagsController from './controllers/feature-flags.js';
 import routeRequiredCapabilities from './routes/required-capabilities.js';
+import ContactSalesLeadsController from './controllers/contact-sales-leads.js';
 
 const uuidRegex = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
 
@@ -225,6 +226,7 @@ async function run(request, context) {
     const tokensController = TokensController(context);
     const plgOnboardingController = PlgOnboardingController(context);
     const imsOrgAccessController = ImsOrgAccessController(context);
+    const contactSalesLeadsController = ContactSalesLeadsController(context);
     const featureFlagsController = FeatureFlagsController(context);
 
     const routeHandlers = getRouteHandlers(
@@ -272,6 +274,7 @@ async function run(request, context) {
       tokensController,
       plgOnboardingController,
       imsOrgAccessController,
+      contactSalesLeadsController,
       featureFlagsController,
     );